+++ src/components/HeaderNav.vue
@@ -110,7 +110,8 @@
 export default {
   data () {
     return {
-      search: ''
+      search: '',
+      userInfo: {}
     }
   },
   computed: {
@@ -122,10 +123,22 @@ export default {
   },
   methods: {
     signout () {
-      window.localStorage.clear()
-      this.$store.commit('CHANGE_LOGIN_STATUS', false)
-      this.$store.commit('GET_SAYHUB_TOKEN', '')
-      // this.$store.commit('GET_USER_INFO', {})
+      this.$confirm('确认登出吗?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        window.localStorage.clear()
+        this.$store.commit('CHANGE_LOGIN_STATUS', false)
+        this.$store.commit('GET_SAYHUB_TOKEN', '')
+        this.$store.commit('GET_USER_INFO', {})
+        this.$router.push({ path: '/' })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消登出'
+        })
+      })
     },
     fetchData () {
       console.log('Search Sayhub')
